refactor(peagen): centralize Status import
