+++ pkgs/standards/peagen/peagen/_utils/_init.py
@@ -12,7 +12,7 @@ import typer
 from peagen.errors import PATNotAllowedError
 from peagen.handlers.init_handler import init_handler
 from peagen.plugins import discover_and_register_plugins
-from peagen.orm.status import Status
+from peagen.transport.jsonrpc_schemas import Status
 from peagen.schemas import TaskCreate
 
 # Allow tests to monkeypatch ``uuid.uuid4`` without affecting the global ``uuid``
