+++ pkgs/standards/peagen/peagen/gateway/db_helpers.py
@@ -8,7 +8,7 @@ from sqlalchemy.dialects.postgresql import insert as pg_insert
 from sqlalchemy.ext.asyncio import AsyncSession
 import sqlalchemy as sa
 
-from peagen.orm.status import Status
+from peagen.transport.jsonrpc_schemas import Status
 from peagen.orm import Base
 from peagen.orm.config.secret import SecretModel
 from peagen.orm.abuse_record import AbuseRecordModel
