+++ pkgs/standards/peagen/peagen/transport/jsonrpc_schemas/status.py
@@ -1,0 +1,7 @@
+"""Status enumeration for RPC schemas."""
+
+from __future__ import annotations
+
+from peagen.orm.status import Status
+
+__all__ = ["Status"]
