+++ pkgs/standards/peagen/peagen/tui/components/filter_bar.py
@@ -6,7 +6,7 @@ from textual.app import ComposeResult
 from textual.containers import Horizontal
 from textual.widgets import Input, Select
 
-from peagen.orm.status import Status
+from peagen.transport.jsonrpc_schemas import Status
 
 
 class FilterBar(Horizontal):
